+++ packages/identity-connector-iota/src/constants/defaultIdentityPkgIds.ts
@@ -4,14 +4,14 @@
 /**
  * Default package IDs for different networks.
  */
-export const DEFAULT_IDENTITY_PKG_IDS = {
+export const DEFAULT_IDENTITY_PKG_IDS: { [id: string]: string } = {
 	/**
 	 * Default package ID for testnet.
 	 */
-	TESTNET: "0x222741bbdff74b42df48a7b4733185e9b24becb8ccfbafe8eac864ab4e4cc555",
+	testnet: "0x222741bbdff74b42df48a7b4733185e9b24becb8ccfbafe8eac864ab4e4cc555",
 
 	/**
 	 * Default package ID for devnet.
 	 */
-	DEVNET: "0x03242ae6b87406bd0eb5d669fbe874ed4003694c0be9c6a9ee7c315e6461a553"
+	devnet: "0x03242ae6b87406bd0eb5d669fbe874ed4003694c0be9c6a9ee7c315e6461a553"
 } as const;