+++ packages/identity-connector-iota/docs/changelog.md
@@ -1,5 +1,61 @@
 # @twin.org/identity-connector-iota - Changelog
 
+## [0.0.1-next.51](https://github.com/twinfoundation/identity/compare/identity-connector-iota-v0.0.1-next.50...identity-connector-iota-v0.0.1-next.51) (2025-06-19)
+
+
+### Features
+
+* support mainnet for alias extraction ([1114e0f](https://github.com/twinfoundation/identity/commit/1114e0f518070ef930d7d6ed73a616cce647206c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @twin.org/identity-models bumped from 0.0.1-next.50 to 0.0.1-next.51
+
+## [0.0.1-next.50](https://github.com/twinfoundation/identity/compare/identity-connector-iota-v0.0.1-next.49...identity-connector-iota-v0.0.1-next.50) (2025-06-19)
+
+
+### Features
+
+* use new generateKid method ([f0fe779](https://github.com/twinfoundation/identity/commit/f0fe779323b675575bb9f80aa74f1957dc57a094))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @twin.org/identity-models bumped from 0.0.1-next.49 to 0.0.1-next.50
+
+## [0.0.1-next.49](https://github.com/twinfoundation/identity/compare/identity-connector-iota-v0.0.1-next.48...identity-connector-iota-v0.0.1-next.49) (2025-06-18)
+
+
+### Miscellaneous Chores
+
+* **identity-connector-iota:** Synchronize repo versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @twin.org/identity-models bumped from 0.0.1-next.48 to 0.0.1-next.49
+
+## [0.0.1-next.48](https://github.com/twinfoundation/identity/compare/identity-connector-iota-v0.0.1-next.47...identity-connector-iota-v0.0.1-next.48) (2025-06-17)
+
+
+### Features
+
+* add mainnet support ([af56a38](https://github.com/twinfoundation/identity/commit/af56a382837896cd4e13d5814c0924c46658b99c))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @twin.org/identity-models bumped from 0.0.1-next.47 to 0.0.1-next.48
+
 ## [0.0.1-next.47](https://github.com/twinfoundation/identity/compare/identity-connector-iota-v0.0.1-next.46...identity-connector-iota-v0.0.1-next.47) (2025-06-12)