+++ packages/identity-models/docs/changelog.md
@@ -1,5 +1,33 @@
 # @twin.org/identity-service-models - Changelog
 
+## [0.0.1-next.51](https://github.com/twinfoundation/identity/compare/identity-models-v0.0.1-next.50...identity-models-v0.0.1-next.51) (2025-06-19)
+
+
+### Miscellaneous Chores
+
+* **identity-models:** Synchronize repo versions
+
+## [0.0.1-next.50](https://github.com/twinfoundation/identity/compare/identity-models-v0.0.1-next.49...identity-models-v0.0.1-next.50) (2025-06-19)
+
+
+### Miscellaneous Chores
+
+* **identity-models:** Synchronize repo versions
+
+## [0.0.1-next.49](https://github.com/twinfoundation/identity/compare/identity-models-v0.0.1-next.48...identity-models-v0.0.1-next.49) (2025-06-18)
+
+
+### Miscellaneous Chores
+
+* **identity-models:** Synchronize repo versions
+
+## [0.0.1-next.48](https://github.com/twinfoundation/identity/compare/identity-models-v0.0.1-next.47...identity-models-v0.0.1-next.48) (2025-06-17)
+
+
+### Miscellaneous Chores
+
+* **identity-models:** Synchronize repo versions
+
 ## [0.0.1-next.47](https://github.com/twinfoundation/identity/compare/identity-models-v0.0.1-next.46...identity-models-v0.0.1-next.47) (2025-06-12)